Clean leveldb locks on shutdown
This change makes the 2i AAE FSM keep track of open leveldb instances
used for temporarily storing index data. When a fold is in progress,
those may remain open for a long time because async workers may have a
long delay before garbage collecting.  The FSM now has a stop command
that will try to close them all nastily and stopping. It will still be
killed by force if that doesn't succeed and the FSM is not responding.
That should be rare, hopefully, as the FSM does very little. The heavy
work is in the workers.