Moving obj update code to riak_kv_index_hashtree
This tries to address some review comments: riak_kv_vnode ended up with
some code related to hashtree updates that seems better hidden inside
the vnode hashtree code.
Cleaned up use of 'Index' in that code.
Change 2i refresh command to take a timeout parameter and avoid
sync_command with infinity timeout, as that's known to not be 100%
reliable and could in rare cases leave this process blocked.