Revert move of object handling to riak_kv_index_hashtree
The refactor I did in that section ended up in code that sends a copy of
the riak object to the hashtree process. That's just wasteful.
Added code to update the hashtree on refresh. The refresh put should
make sure the hashtree information is up to date to ensure convergence
of these repairs