Adding missing skipped_read_repairs stats
These stats are updated by the code when it decides that a read repair
should be performed, but it is not a good idea to do it at the moment.
However, the plumbing to implement that stat was missing. This adds it
so that you have skipped_read_repairs and skipped_read_repairs_total
stats through the http interface and riak-admin status.