Add Bitcask backend upgrade procedure
At startup, the backend looks at the Bitcask app version and a version
written in the data directory, if any.  If the pair require an upgrade,
an upgrade file is written to the directory and a regular check is
scheduled to determine when the upgrade is finished. When it is, the new
version is written.

Upgrading to Bitcask 1.7.0 schedules a merge for all existing files by
writing a merge.txt with the list of files in the data directory.
Bitcask's need to merge (starting in 1.7.0) will read that and use the
list for merging, deleting the file when all files have been merged.

This also adds a cuttlefish configuration parameter for the Bitcask
merge check interval.