Add support for KV updates socket
If riak_kv_events_manager:post_hook/1 is installed as a post-commit hook,
KV updates are sent as protobuf messages over a TCP socket to registered
subscribers. Subscribers connect to the riak_api.pbevents port. Avert your
eyes, there are a number of hacks here, but remember it's just a prototype.