Converting stop_fold exception breaks vnode worker
If the riak_kv_worker sees a stop_fold exception, it will stop and do no
further work. This code was blocking that exception and handling it in
the eleveldb by return the value of the accumulator *before* the
exception. This is wrong. The kv fold logic has already processed that
accumulator and sent the results back. This causes some extra values to
be appended to the results. In the case of paginated 2i queries that use
list:merge to do a merge sort of the results, the extra values made the
partial lists not be in sorted order, and list:merge in turn returns
results that are in the incorrect order. This causes some results to be
moved to the end and be left out of the final merged result.