Hotfix to canonicalize hashes optimized for HTTP interface stored objects.
If the object metadata dict has variations in the order of entries it
causes the hashes to differ.  This hotfix tries to get it into a likely
format for HTTP written objects and if that fails, just sorts the metadata
so it will hash the same.

KV AAE will detect the difference in hashes and issue a read repair
which will not do anything as the objects are equivalent according to
syntactic_merge.

However the riak_kv_exchange_fsm triggers reindexing.  There is a race
in riak_kv_exchange_fsm the riak_client request does not wait for the
read repair to be issued, but in this case it is fine as the data will
not change and we just desire the rehash behavior in the AAE tree anyway.